refactor: 💡 추가 PR 에 조건들 추가
