+++ src/main/java/store/cookshoong/www/cookshoongbackend/address/model/request/CreateAccountAddressRequestDto.java
@@ -1,4 +1,4 @@
-package store.cookshoong.www.cookshoongbackend.address.dto.request;
+package store.cookshoong.www.cookshoongbackend.address.model.request;
 
 import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
@@ -11,12 +11,12 @@ import lombok.Getter;
  * @author jeongjewan
  * @since 2023.07.04
  */
-
 @Getter
 @AllArgsConstructor
 public class CreateAccountAddressRequestDto {
 
     private String alias;
+
     private String mainPlace;
 
     private String detailPlace;
