+++ src/main/java/store/cookshoong/www/cookshoongbackend/address/model/response/AddressResponseDto.java
@@ -1,4 +1,4 @@
-package store.cookshoong.www.cookshoongbackend.address.dto.response;
+package store.cookshoong.www.cookshoongbackend.address.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,7 +10,6 @@ import lombok.Getter;
  * @author jeongjewan
  * @since 2023.07.04
  */
-
 @Getter
 @AllArgsConstructor
 public class AddressResponseDto {
