+++ src/main/java/store/cookshoong/www/cookshoongbackend/address/repository/accountaddress/AccountAddressRepositoryCustom.java
@@ -2,15 +2,14 @@ package store.cookshoong.www.cookshoongbackend.address.repository.accountaddress
 
 import java.util.List;
 import org.springframework.data.repository.NoRepositoryBean;
-import store.cookshoong.www.cookshoongbackend.address.dto.response.MainAddressResponseDto;
+import store.cookshoong.www.cookshoongbackend.address.model.response.MainAddressResponseDto;
 
 /**
- * {설명을 작성해주세요}.
+ * 회원과 주소 repository.
  *
  * @author jeongjewan
  * @since 2023.07.04
  */
-
 @NoRepositoryBean
 public interface AccountAddressRepositoryCustom {
